requirements.txt
fastapi
uvicorn[standard]

pytest
httpx

streamlit

langchain
google-generativeai

langchain-google-genai
python-dotenv


